refactor(auth): reorganize imports in main.js for better clarity and structure
