+++ backend/cloud/controllers/auth/validate-user.js
