+++ backend/cloud/main.js
@@ -1,8 +1,8 @@
-require("./controllers/auth/validateUser");
 require("./controllers/auth/login");
+require("./controllers/auth/update-phone");
+require("./controllers/auth/validate-user");
 require("./controllers/auth/create-account");
 require("./controllers/auth/is-subscription-active");
-require("./controllers/auth/update-phone");
 
 require("./controllers/payment");
 require("./controllers/general/create-company");
